+++ src/js/components/drawer.js
@@ -46,7 +46,7 @@ export class Drawer extends Dialog {
         this.opts.className,
       ].filter(Boolean).join(' '),
       id: this.id,
-      ...(this.opts.title ? { 'aria-labelledby': titleId } : {}),
+      ...(this.opts.title ? { 'aria-labelledby': titleId } : this.opts.text ? { 'aria-label': this.opts.text } : {}),
     }, [this.panel]);
 
     this.body = this.panel.querySelector('.tuc-drawer__body');